requirements.txt
django==2.0.5
djangorestframework==3.10.2
psycopg2==2.8.3
djangorestframework-jwt